+++ app/src/main/res/values-v21/themes_dark.xml
@@ -20,7 +20,8 @@
         <item name="dialogTheme">@style/DialogThemeDark</item>
 
         <item name="colorControlHighlight">@color/dark_ripple</item>
-
+         <!--        snackbar theming-->
+        <item name="snackbarStyle">@style/Widget.App.SnackbarDefaultDarkTheme</item>
         <item name="android:ambientShadowAlpha">0.4</item>
         <item name="android:spotShadowAlpha">0.4</item>
         <item name="android:outlineAmbientShadowColor" tools:targetApi="p">@color/dark_glow</item>
@@ -32,6 +33,9 @@
         <item name="drawerClosedBackgroundColor">@color/dark_background</item>
     </style>
 
+    <style name="Widget.App.SnackbarDefaultDarkTheme" parent="" >
+        <item name="android:background">@color/dark_secondary</item>
+    </style>
     <style name="AlarmAlertFullScreenDarkTheme" parent="Theme.AppCompat.NoActionBar">
         <item name="android:windowNoTitle">true</item>
         <item name="colorAccent">@color/dark_secondary</item>
